:recycle: refactor: Header 하위 컴포넌트 폴더 분리 및 상태/핸들러 구조 정리
